Added in a few more of the types and adjusted some of the naming conventions of dataFlattener.ts, now named getFlattenedRegionCodes.ts.
